requirements.txt
fastapi
uvicorn
pydantic
requests
pdfplumber
tiktoken
faiss-cpu
openai
python-dotenv
python-docx 
beautifulsoup4
pymupdf
langchain-mcp-adapters
langchain-community
mcp
langchain-openai
langchain
langgraph
httpx[http2]
openpyxl
python-calamine
python-pptx
huggingface-hub 
httpx
python-pptx
easyocr
Pillow
pandas
rapidocr-onnxruntime
memory_profiler
